+++ astropy/tests/__init__.py
@@ -4,3 +4,6 @@ This package contains utilities to run the astropy test suite, tools
 for writing tests, and general tests that are not associated with a
 particular package.
 """
+
+RUNNING = False
+"""Flag that indicates if our unit tests are running."""